LayoutTests/accessibility/mac/aria-tree-with-presentation-role-expected.txt

Release Notes
Using Search
Power Search
Common Info Area
This tests that aria trees are correctly exposed when there is role='presentation' presented.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


AXRole: AXOutline AXValue:
  AXRole: AXRow level: 0
    AXRole: AXStaticText AXValue: Release Notes
  AXRole: AXRow level: 0
    AXRole: AXStaticText AXValue: Using Search
  AXRole: AXRow level: 1
    AXRole: AXStaticText AXValue: Power Search
  AXRole: AXRow level: 0
    AXRole: AXStaticText AXValue: Common Info Area
PASS successfullyParsed is true

TEST COMPLETE
